requirements/test.txt
-r common.txt

mock==4.0.2
pytest-cov==2.10.1
pytest-django==3.9.0
requests-mock==1.8.0
